feat: компонент AddToFavorite
