+++ src/entities/AddToFavorite/index.ts
@@ -1,0 +1,2 @@
+'use client';
+export { AddToFavorite } from './ui';
