+++ src/entities/AddToFavorite/styled.ts
@@ -1,0 +1,7 @@
+import { styled } from 'styled-components';
+
+export const StButton = styled.button`
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+`;
