+++ src/features/FavoriteItem/typing.ts
