+++ src/widgets/ProductCard/index.ts
@@ -1,1 +1,2 @@
+'use client';
 export { ProductCard } from './ui';
